# pkg/swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Chat-Server API",
    "version": "1.0.0",
    "contact": {
      "name": "Konstantin Ivanov",
      "email": "mrkgeniy@gmail.com"
    }
  },
  "tags": [
    {
      "name": "ChatServerV1"
    }
  ],
  "host": "localhost:50662",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/chat_server/v1": {
      "delete": {
        "operationId": "ChatServerV1_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "ChatServerV1"
        ]
      },
      "post": {
        "operationId": "ChatServerV1_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/chat_server_v1CreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/chat_server_v1CreateRequest"
            }
          }
        ],
        "tags": [
          "ChatServerV1"
        ]
      }
    },
    "/chat_server/v1/send-message": {
      "post": {
        "operationId": "ChatServerV1_SendMessage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/chat_server_v1SendMessageRequest"
            }
          }
        ],
        "tags": [
          "ChatServerV1"
        ]
      }
    }
  },
  "definitions": {
    "chat_server_v1CreateRequest": {
      "type": "object",
      "properties": {
        "usernames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "chat_server_v1CreateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "chat_server_v1SendMessageRequest": {
      "type": "object",
      "properties": {
        "chatId": {
          "type": "string",
          "format": "int64"
        },
        "from": {
          "type": "string",
          "title": "TODO ??? int or string [(validate.rules).int64.gt = 0];"
        },
        "text": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}